Withdraw the "--buildcmd" argument from `mdm release`.
There are three pieces of context needed for a buildcmd to make any sense: the directory of the project, the directory to put the built artifacts into, and the version number.  The previous implementation had only accounted for the second and third of those; pending a new design plan that makes all of these available more elegantly, this option is removed.

More importantly, having the build process either be distinctly before or be the invoker of the release process appears to work really quite well.