Begin defining options for `mdm depend`.
This is a first draft and there are several things about it that make me a bit dissatisified.  In particular, the fact that it ended up with two slightly disparate defintions for "name" rather concerns me.  I may revise this heavily in the future after doing a little more research into what kind of metadata can be tagged to a submodule.