Refactor some code out into other module files.
The abstraction is still leaky, certainly, since this really is essentially an overgrown shell script expressed in python; that's not going to change as long as this tool is based on pbs (and in turn pbs's carefree usage of cwd doesn't make things any easier).  Still, this is a half step cleaner.