Implement most of `mdm update`.
Making fields package-writable in MdmModule when the situation demands it, and not really having any shame about it.  This is going to make it very easy to desync our in memory state from reality.  At the moment I don't care, because every one of this commands is pretty short-lived, but if it becomes a problem then building some more sensible cache refreshing behavior into MdmModule might be wise.  ...Or, I quit being silly and just put all those methods in Plumbing into MdmModule.  But I'm going to run with the pattern of fairly direct translation from python for a while yet, and find reasonable places to break things up later.