Remove .jar suffix from product.
Most linux envirnments seem to be happy to accept a zip file of this form as a java application and just roll with it.  Windows... well.  It's really sad how java is still the most realistic way to ship something cross-platform, but then that last yard with launching is such a clusterf--k.