Create basic infrastructure for unit tests, in particular to deal with temporary working directories.
This is less than ideal in some ways -- in particular, there's going to be a wholllle lot of instances of new File(.).getCanonicalPath() in test methods to try to pretend we have a real valid cwd.  Integration tests in the future will try to address coverage of this, but I think that will require doing full forking per test, which isn't a great option for unit tests.