Add parse and validate methods to MdmCommand interface.
These are already implemented in some commands, and will be in the rest in the future.  Now that these functions are defined for all MdmCommand, I can call them from the main method in Mdm, and remove the placeholder calls at the top of command already using the pattern, which makes it easier to compose commands programmatically in tests without having to fake args objects.