create test fixture package and start moving relevant code to it.
the reliance on MdmReleaseInitCmd to set up other tests is a little uncomfortable but also more or less inevitable.

i'm very unhappy about the visibility bump some random fields got in this change... it's embarassing how much of this project is still reeling from the pass-maps-everywhere python design port.  this should go away when we get to proper builder interfaces for command config.