refactor the main class so that there's a reasonably well exposed method that acts like the cli but without execing and without halting the jvm on errors.
this is essentially meant for testing access.  an alternative would be to drive the separation between command parameters and implementation details... but really, i have no reason to be interested in having multiple implementations of params provider interfaces or etc; at this point the practical utility of that would be somewhere around using mocks, and there really isn't any call for that.