don't show the password on the screen as it's being entered.
shame the java Console api admitted enough platform knowledge to get us '-echo' mode on the terminal, but not enough to give us a quick call to flip to uncooked input buffering.

and yes, incidentally, I know that the readPassword function returns a char[] for a reason -- yes, I know it's considered advisable to zero the password memory after use, and yes, I know it's quite hard to do that once you've made a string in java.  But I'm A) not getting in an argument with jsch's interfaces right now, and B) firmly of the opinion that if you have worries about your process memory being inspected by an adversary, your pooch is already pretty screwed and it's time to generate new keys and reevaluate your lifestyle.