update to stance on transitive dependency resolution.
it's of course still true that a transitive dependency resolution step at runtime will *never* be necessary for someone to *use* a project that has deps spec'd with mdm.

and dependency resolvers have some downsides: they historically do seem to have a tendency to bloat projects (you can't trust your dependency's dependency to not specify fucking junit as a compile dependency, and similarly there are string processing libraries on maven right this moment that rely on both scala-lang *and* jruby in the same breath and no visible sense of irony), and dependency resolve adoption actually tends raise the complexity level of an ecosystem(more tiny jars does not help, if i wanted a smaller binary i'd use a real tree shaker like proguard because That's What They're For).

but all that said, i'm open the idea a transitive dependency resolver can possibly be useful.  or at the very least, as long as they produce firmly pegged and version controlled lists of final, resolved dependencies, then they can't spectacularly damage the ability of a project to ship, which is good enough.