thinking towards features for allowing updates and alters to force overwrite of a dirty working tree.
current lack of explicit support for this doesn't block any flows, because one can always rm-rf the offending directory and do a fresh update, or interact with it like a normal git repo via checkout, reset, etc.  however, easier first-class support for this does seem to be a fairly popular request.