junit output: i do not want to see the stdout or stderr of tests unless something has gone horribly wrong.
sadly, this only works *sometimes*.  it's not nearly as smart as what `go test` does.  seems like subprocesses spawned by josh still see my tty.  i don't know if this is junit's fault, or because something deep in the java subprocess forking model is fucked... based on prior experiences, probably the latter actually.  no matter, they're going to rewrite it AGAIN in the next version of java, and completely re-fuck all the sematics of filedescriptors in a unique and exciting way...